[Release] merge branch 'development' into 'implementation'

+++ back-end/hospital-api/src/main/java/net/pladema/user/controller/BackofficeUserRolesStore.java
@@ -5,10 +5,12 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
+import net.pladema.establishment.repository.InstitutionalGroupUserRepository;
 import net.pladema.permissions.repository.RoleRepository;
 
 import net.pladema.permissions.repository.entity.Role;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -40,6 +42,7 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 	private final ProfessionalProfessionRepository professionalProfessionRepository;
 	private final BackofficeRolesFilter backofficeRolesFilter;
 	private final RoleRepository roleRepository;
+	private final InstitutionalGroupUserRepository institutionalGroupUserRepository;
 
 
 	@Override
@@ -80,10 +83,22 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 		validate(entity);
 		if (entity.getInstitutionId() == null)
 			entity.setInstitutionId(-1);
-		return userRoleRepository
-				.getUserRoleIfIsDeleted(entity.getUserId(), entity.getRoleId(), entity.getInstitutionId())
-				.map(userRoleRepository::reactivate)
-				.orElseGet( () -> userRoleRepository.save(entity));
+
+		try {
+			Optional<UserRole> userRoleAlreadyExist = userRoleRepository.findByRoleInstitutionAndUserId(
+					entity.getUserId(),
+					entity.getRoleId(),
+					entity.getInstitutionId()
+			);
+			if (userRoleAlreadyExist.isPresent())
+				throw new DataIntegrityViolationException("El usuario ya cuenta con ese rol en la institución");
+			return userRoleRepository
+					.getUserRoleIfIsDeleted(entity.getUserId(), entity.getRoleId(), entity.getInstitutionId())
+					.map(userRoleRepository::reactivate)
+					.orElseGet( () -> userRoleRepository.save(entity));
+		} catch (DataIntegrityViolationException e) {
+			throw new BackofficeUserException(BackofficeUserExceptionEnum.ROLE_ALREADY_ASSIGNED, "El usuario ya cuenta con ese rol en la institución");
+		}
 	}
 
 	private void validate(UserRole userRole) {
@@ -102,10 +117,9 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_INVALID_ROLE, "El usuario creado no puede tener el siguiente rol: ROOT");
 		if (ERoleLevel.LEVEL1.equals(ERole.map(userRole.getRoleId()).getLevel()) && !assignedInstitution(userRole))
 			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_INVALID_ROLE, "El tipo de rol requiere definir una institución ");
-		if (ERoleLevel.LEVEL1.equals(ERole.map(userRole.getRoleId()).getLevel()) && !assignedInstitution(userRole))
-			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_INVALID_ROLE, "El tipo de rol requiere definir una institución ");
 		if (ERoleLevel.LEVEL0.equals(ERole.map(userRole.getRoleId()).getLevel()) && assignedInstitution(userRole))
 			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_INVALID_ROLE, "El tipo de rol no debe asociarse a una institución ");
+		checkManagerRole(userRole);
 		return user;
 	}
 
@@ -139,7 +153,8 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 				ERole.PROFESIONAL_DE_SALUD.getId().equals(roleId) ||
 				ERole.ESPECIALISTA_EN_ODONTOLOGIA.getId().equals(roleId) ||
 				ERole.VIRTUAL_CONSULTATION_PROFESSIONAL.getId().equals(roleId) ||
-				ERole.VIRTUAL_CONSULTATION_RESPONSIBLE.getId().equals(roleId);
+				ERole.VIRTUAL_CONSULTATION_RESPONSIBLE.getId().equals(roleId) ||
+				ERole.ABORDAJE_VIOLENCIAS.getId().equals(roleId);
 	}
 
 	@Override
@@ -156,6 +171,8 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 			if(isROOTUser(user.getUsername()))
 				throw new BackofficeUserException(BackofficeUserExceptionEnum.ROOT_LOST_PERMISSION, "El admin no puede perder el rol: ROOT");
 		}
+		if (isManager(userRole) && institutionalGroupUserRepository.existsByUserId(userRole.getUserId()))
+			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_BELONGS_TO_GROUP, "El usuario no puede perder el rol ya que se encuentra asociado a grupos institucionales");
 		userRoleRepository.deleteById(id);
 	}
 
@@ -172,6 +189,21 @@ public class BackofficeUserRolesStore implements BackofficeStore<UserRole, Long>
 		return StreamSupport
 				.stream(roles.spliterator(), false)
 				.collect(Collectors.toList());
+	}
+
+	private void checkManagerRole(UserRole userRole){
+		if (!isManager(userRole))
+			return;
+		List<Short> rolesIds = userRoleRepository.findByUserId(userRole.getUserId()).stream().map(UserRole::getRoleId).collect(Collectors.toList());
+		if (rolesIds.contains(ERole.GESTOR_DE_ACCESO_DE_DOMINIO.getId()) || rolesIds.contains(ERole.GESTOR_DE_ACCESO_REGIONAL.getId()) || rolesIds.contains(ERole.GESTOR_DE_ACCESO_LOCAL.getId()))
+			throw new BackofficeUserException(BackofficeUserExceptionEnum.USER_INVALID_ROLE, "El usuario ya cuenta con un rol de tipo Gestor");
+	}
+
+	private boolean isManager(UserRole role){
+		Short roleId = role.getRoleId();
+		return roleId.equals(ERole.GESTOR_DE_ACCESO_DE_DOMINIO.getId()) ||
+				roleId.equals(ERole.GESTOR_DE_ACCESO_REGIONAL.getId()) ||
+				roleId.equals(ERole.GESTOR_DE_ACCESO_LOCAL.getId());
 	}
 
 }
